fix build

+++ build_vpn.sh
@@ -8,10 +8,22 @@ fi
 pushd jni/ShadowVPN || exit 1
 ./autogen.sh || exit 1
 
-dist-build/android-arm.sh || exit 1
-dist-build/android-armv7.sh || exit 1
-dist-build/android-mips.sh || exit 1
-dist-build/android-x86.sh || exit 1
+if [ ! -f shadowvpn-android-arm/lib/libshadowvpn.a ]; then
+    dist-build/android-arm.sh || exit 1
+fi
+if [ ! -f shadowvpn-android-armv7/lib/libshadowvpn.a ]; then
+    dist-build/android-armv7.sh || exit 1
+fi
+if [ ! -f shadowvpn-android-mips/lib/libshadowvpn.a ]; then
+    dist-build/android-mips.sh || exit 1
+fi
+if [ ! -f shadowvpn-android-x86/lib/libshadowvpn.a ]; then
+    dist-build/android-x86.sh || exit 1
+fi
+
+popd
+
+pushd jni
 
 install -d armeabi
 install -d armeabi-v7a
@@ -28,9 +40,7 @@ install ShadowVPN/shadowvpn-android-armv7/lib/libshadowvpn.a armeabi-v7a
 install ShadowVPN/shadowvpn-android-mips/lib/libshadowvpn.a mips
 install ShadowVPN/shadowvpn-android-x86/lib/libshadowvpn.a x86
 
-popd
 
-pushd jni
 $ANDROID_NDK_HOME/ndk-build || exit 1
 popd
 
@@ -41,6 +51,6 @@ install -d app/src/main/jniLibs/x86
 
 install libs/armeabi/libvpn.so app/src/main/jniLibs/armeabi
 install libs/armeabi-v7a/libvpn.so app/src/main/jniLibs/armeabi-v7a
-install libs/armeabi-mipa/libvpn.so app/src/main/jniLibs/mips
-install libs/armeabi/x86.so app/src/main/jniLibs/x86
+install libs/mips/libvpn.so app/src/main/jniLibs/mips
+install libs/x86/libvpn.so app/src/main/jniLibs/x86
